+++ src/plugins/svgo.js
@@ -1,7 +1,0 @@
-import SVGO from 'svgo'
-
-export default async (code, opts) => {
-  const svgo = new SVGO(opts)
-  const { data } = await svgo.optimize(code)
-  return data
-}