+++ src/plugins/transform.js
@@ -1,1 +1,0 @@
-export default (code, opts, state) => opts.transform(code, state)