+++ src/transforms/rename.js
@@ -1,3 +1,0 @@
-import { upperFirst, camelCase } from 'lodash'
-
-export const pascalCase = str => upperFirst(camelCase(str))